requirements.txt
loguru==0.7.2
python-dotenv==1.0.1
requests==2.31.0
aiogram==2.23.1
currencyapicom==0.1.1